refactor: changed animation, ui elements locations added constraints for fingerprint button, renamed files
